@@ -21,7 +21,7 @@
     "build:main": "node scripts/build-main.js",
     "build:logger": "node scripts/build-logger.js",
     "lint": "eslint src test",
-    "test": "npm run lint && npm run test:unit && npm run test:ssr && npm run test:e2e",
+    "test": "npm run lint && npm run test:unit && npm run test:ssr && npm run test:types && npm run test:e2e",
     "test:unit": "jasmine JASMINE_CONFIG_PATH=test/unit/jasmine.json",
     "test:e2e": "node test/e2e/runner.js",
     "test:ssr": "cross-env VUE_ENV=server jasmine JASMINE_CONFIG_PATH=test/unit/jasmine.json",
@@ -49,6 +49,7 @@
     "@rollup/plugin-commonjs": "^11.1.0",
     "@rollup/plugin-node-resolve": "^7.1.3",
     "@rollup/plugin-replace": "^2.3.1",
+    "@types/node": "^13.13.0",
     "@vue/compiler-sfc": "3.0.0-beta.2",
     "babel-core": "^6.22.1",
     "babel-loader": "^7.1.2",
@@ -74,7 +75,7 @@
     "rollup": "^2.6.1",
     "rollup-plugin-terser": "^5.3.0",
     "todomvc-app-css": "^2.1.0",
-    "typescript": "^3.7.2",
+    "typescript": "^3.8.3",
     "vue": "3.0.0-beta.2",
     "vue-loader": "16.0.0-alpha.3",
     "vue-template-compiler": "^2.5.22",
