@@ -1,4 +1,4 @@
-import Vue from "vue";
+import { createApp } from "vue";
 import * as Vuex from "../index";
 
 const store = new Vuex.Store({
@@ -7,8 +7,10 @@ const store = new Vuex.Store({
   }
 });
 
-const vm = new Vue({
+const app = createApp({
   store
 });
+
+const vm = app.mount('#el')
 
 vm.$store.state.value;
