@@ -1,4 +1,4 @@
-import Vue from 'vue';
+import { ComponentPublicInstance } from 'vue';
 import { Dispatch, Commit } from './index';
 
 type Computed = () => any;
@@ -6,7 +6,7 @@ type InlineComputed<T extends Function> = T extends (...args: any[]) => infer R 
 type MutationMethod = (...args: any[]) => void;
 type ActionMethod = (...args: any[]) => Promise<any>;
 type InlineMethod<T extends (fn: any, ...args: any[]) => any> = T extends (fn: any, ...args: infer Args) => infer R ? (...args: Args) => R : never
-type CustomVue = Vue & Record<string, any>;
+type CustomVue = ComponentPublicInstance & Record<string, any>;
 
 interface Mapper<R> {
   <Key extends string>(map: Key[]): { [K in Key]: R };
